Migrate the RTT refinement service onto LatestRepository
Replaces both private Maps, readUnexpiredEntry, pruneExpiredEntries,
pruneExpired, the RateLimiter field and its construction, and the two entry
type aliases with two LatestRepository instances. claimWork is now read ->
readOrAcceptAt -> acceptAt, still synchronous and still sampling one instant
that it threads through every judgement.

All six existing service tests pass unchanged, including the two that pin the
amortized scan and the read-time expiry, so behaviour is preserved rather than
re-specified.

readOrAcceptAt no longer demands a placeholder value alongside create.

The worked example test goes: it existed to show the shape against a real
caller before migrating, and the real caller now does it.

One deliberate difference: eviction budget is per repository rather than one
budget covering both maps, so a claim that writes to both can now scan twice
where it previously scanned once. Both scans are still rate limited and the
scan is still off the per-claim path.
