+++ packages/shared-server/rallar-system/rtc-topology/topic/rtc-rtt-refinement-service.ts
@@ -1,7 +1,7 @@
 import type { RttMeasurementInfo } from '@shared/api/api-config.ts';
 import { predictedRttMs, type VivaldiNodeData } from '@shared-graph/graph/vivaldi.ts';
 
-import { RateLimiter } from '@shared/resilience/Resilience.ts';
+import { LatestRepository } from '@shared/cache/LatestRepository.ts';
 
 import type { RtcRttRefinementGate } from './rtc-rtt-refinement-gate.ts';
 
@@ -31,71 +31,56 @@ export interface ClaimRtcRttRefinementWorkInput {
   readonly expireAtEpochMs: number;
 }
 
-type ExpiringObservation = Readonly<{
-  predictedDeltaMs: number;
-  expireAtEpochMs: number;
-}>;
-
-type ExpiringDecision = Readonly<{
-  claimed: boolean;
-  expireAtEpochMs: number;
-}>;
-
 /** Coordinates stable refinement decisions across durable work retries. */
 export class RtcRttRefinementService {
   private readonly dependencies: RtcRttRefinementServiceDependencies;
-  private readonly pruneRateLimiter: RateLimiter;
-  private readonly observations = new Map<string, ExpiringObservation>();
-  private readonly decisions = new Map<string, ExpiringDecision>();
+  private readonly observations: LatestRepository<string, number>;
+  private readonly decisions: LatestRepository<string, boolean>;
 
   constructor(dependencies: RtcRttRefinementServiceDependencies) {
     this.dependencies = dependencies;
     const pruneLimit = dependencies.pruneLimit ?? DEFAULT_RTT_REFINEMENT_PRUNE_LIMIT;
-    this.pruneRateLimiter = RateLimiter.initWithTs(
-      pruneLimit.windowMs,
-      pruneLimit.maxPerWindow,
-      dependencies.nowEpochMs(),
-    );
+    this.observations = new LatestRepository({
+      evictWindowMs: pruneLimit.windowMs,
+      evictsPerWindow: pruneLimit.maxPerWindow,
+    });
+    this.decisions = new LatestRepository({
+      evictWindowMs: pruneLimit.windowMs,
+      evictsPerWindow: pruneLimit.maxPerWindow,
+    });
   }
 
   claimWork(input: ClaimRtcRttRefinementWorkInput): boolean {
     const nowEpochMs = this.dependencies.nowEpochMs();
-    if (this.pruneRateLimiter.allowAt(nowEpochMs)) {
-      this.pruneExpired(nowEpochMs);
-    }
-    const existingDecision = readUnexpiredEntry(this.decisions, input.workId, nowEpochMs);
-    if (existingDecision) return existingDecision.claimed;
+    const existingDecision = this.decisions.readAt(input.workId, nowEpochMs);
+    if (existingDecision !== undefined) return existingDecision;
 
-    const predictedDeltaMs = this.readOrObserveDelta(input, nowEpochMs);
+    const predictedDeltaMs = this.observations.readOrAcceptAt({
+      key: input.observationId,
+      nowEpochMs,
+      expireAtEpochMs: input.expireAtEpochMs,
+      create: () => this.observeAndMeasurePredictedDelta(input.rtt),
+    });
+
     const claimed = this.dependencies.gate.claimRefinement({
       groupKey: input.groupKey,
       predictedDeltaMs,
       nowEpochMs,
     });
-    this.decisions.set(input.workId, {
-      claimed,
+    this.decisions.acceptAt({
+      key: input.workId,
+      value: claimed,
+      nowEpochMs,
       expireAtEpochMs: input.expireAtEpochMs,
     });
     return claimed;
   }
 
   readRetainedEntryCounts(): Readonly<{ observations: number; decisions: number }> {
-    return { observations: this.observations.size, decisions: this.decisions.size };
-  }
-
-  private readOrObserveDelta(
-    input: ClaimRtcRttRefinementWorkInput,
-    nowEpochMs: number,
-  ): number {
-    const existing = readUnexpiredEntry(this.observations, input.observationId, nowEpochMs);
-    if (existing) return existing.predictedDeltaMs;
-
-    const predictedDeltaMs = this.observeAndMeasurePredictedDelta(input.rtt);
-    this.observations.set(input.observationId, {
-      predictedDeltaMs,
-      expireAtEpochMs: input.expireAtEpochMs,
-    });
-    return predictedDeltaMs;
+    return {
+      observations: this.observations.size(),
+      decisions: this.decisions.size(),
+    };
   }
 
   private observeAndMeasurePredictedDelta(rtt: RttMeasurementInfo): number {
@@ -106,11 +91,6 @@ export class RtcRttRefinementService {
       ? Number.POSITIVE_INFINITY
       : Math.abs(predictedAfter - predictedBefore);
   }
-
-  private pruneExpired(nowEpochMs: number): void {
-    pruneExpiredEntries(this.observations, nowEpochMs);
-    pruneExpiredEntries(this.decisions, nowEpochMs);
-  }
 }
 
 function readPredictedPairRttMs(
@@ -120,27 +100,4 @@ function readPredictedPairRttMs(
   const fromNode = nodes.get(rtt.sessionIdFrom);
   const toNode = nodes.get(rtt.sessionIdTo);
   return fromNode && toNode ? predictedRttMs(fromNode, toNode) : undefined;
-}
-
-function readUnexpiredEntry<T extends Readonly<{ expireAtEpochMs: number }>>(
-  entries: Map<string, T>,
-  key: string,
-  nowEpochMs: number,
-): T | undefined {
-  const entry = entries.get(key);
-  if (entry === undefined) return undefined;
-  if (entry.expireAtEpochMs <= nowEpochMs) {
-    entries.delete(key);
-    return undefined;
-  }
-  return entry;
-}
-
-function pruneExpiredEntries<T extends Readonly<{ expireAtEpochMs: number }>>(
-  entries: Map<string, T>,
-  nowEpochMs: number,
-): void {
-  for (const [key, value] of entries) {
-    if (value.expireAtEpochMs <= nowEpochMs) entries.delete(key);
-  }
 }
